PythonDemo/pyvisaDemo/requirements.txt
pyvisa>=1.13.0
pyvisa-py>=0.7.0
PyYAML>=6.0
numpy>=1.24.0
pandas>=2.0.0
openpyxl>=3.1.0
matplotlib>=3.7.0
ttkbootstrap>=1.10.0
Jinja2>=3.1.0
